requirements.txt
albumentations==0.5.2
annotated-types==0.6.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.8
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cmake==3.25.0
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.1.1
cycler==0.12.1
dacite==1.8.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.4
executing==0.10.0
fastjsonschema==2.16.3
filelock==3.9.0
flatbuffers==23.5.26
flopth==0.1.3
fonttools==4.47.0
fqdn==1.5.1
fsspec==2023.12.2
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.40
htmlmin==0.1.12
huggingface-hub==0.20.1
idna==3.4
ImageHash==4.3.1
imageio==2.33.1
imgaug==0.4.0
importlib-metadata==7.0.1
importlib-resources==5.12.0
ipykernel==6.28.0
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lazy_loader==0.3
lit==15.0.7
llvmlite==0.41.1
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
mistune==2.0.5
mpmath==1.3.0
multimethod==1.10
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nest-asyncio==1.5.5
networkx==3.0
notebook==6.5.3
notebook_shim==0.2.2
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnx==1.14.0
onnxruntime==1.15.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.0.3
pandas-profiling==3.6.6
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.6
pexpect==4.8.0
phik==0.12.4
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
prometheus-client==0.16.0
prompt-toolkit==3.0.30
protobuf==4.23.3
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycocotools==2.0.6
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.13.0
pyparsing==3.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-triton==0.0.1
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
qudida==0.0.4
regex==2023.12.25
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
safetensors==0.4.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
sentry-sdk==1.39.1
setproctitle==1.3.3
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4
stack-data==0.4.0
statsmodels==0.14.1
sympy==1.12
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.12
tinycss2==1.2.1
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
torchxrayvision==1.2.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
triton==2.0.0
typeguard==4.1.5
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.2.0
urllib3==1.26.13
visions==0.7.5
wandb==0.16.1
wcwidth==0.2.12
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.9
wordcloud==1.9.3
ydata-profiling==4.6.4
zipp==3.17.0
